Change boot to watch only no operation
- Split loader from launcher
- Change timeout and add new attribute 'operate' to config if it should operate for you or notify
- Add logging levels to improve visualization
- Fix symbol to a valid one.
