+++ configuration.yml
@@ -18,4 +18,7 @@ ta_summary:
   takeprofit: 0.02 #type=float, price to reach above (for long) or below (for short) entry price to take profit, 0.2 means take profit 20% above or below entry price 
   stoploss: 0.02 #type=float, price to reach below (for long) or above (for short) entry price to stop loss, 0.1 means stop loss 10% below or above entry price
   interval: '1m' #type=string, available interval between 2 TA summary signals: "1m" for 1 minute, "5m" for 5 minutes, "15m" for 15 minutes, "30m" for 30 minutes, "1h" for 1 hour, "2h" for 2 hours, "4h" for 4 hours, "1d" for 1 day, "1W" for 1 week, "1M" for 1 month
-  timeout: 60 #type=int, min=1, number of minutes the trading strat will be running2
+  timeout: 6000 #type=int, min=1, number of minutes the trading strat will be running2
+
+## choose if it will do operations or just log them
+operate: False
