+++ launcher.py
@@ -1,17 +1,5 @@
 from strategies.ta_summary import TAS
-import os
-import yaml
-
-
-# To load Configuration.yml file
-def load_yaml(file):
-    with open(file) as file:
-        load = yaml.load(file, Loader=yaml.FullLoader)
-    return load
-
-yaml_file = load_yaml(os.path.join(os.path.dirname(__file__), "configuration.yml"))
-
-lnm_options = yaml_file["lnm_credentials"]
+from config_loader import yaml_file, lnm_options
 
 # call the bot you have choosed in configuration file
 def bot():
